Create and Implement AuthenticationPresenter
