+++ app/src/main/java/edu/byu/cs/tweeter/client/presenter/view/AuthenticationView.java
@@ -1,4 +1,4 @@
-package edu.byu.cs.tweeter.client.view;
+package edu.byu.cs.tweeter.client.presenter.view;
 
 import edu.byu.cs.tweeter.model.domain.User;
