+++ app/src/main/java/edu/byu/cs/tweeter/client/presenter/view/View.java
@@ -1,4 +1,4 @@
-package edu.byu.cs.tweeter.client.view;
+package edu.byu.cs.tweeter.client.presenter.view;
 
 public interface View {
     void displayMessage(String message);
