+++ app/src/main/java/edu/byu/cs/tweeter/client/view/login/RegisterFragment.java
@@ -23,7 +23,7 @@ import java.io.ByteArrayOutputStream;
 import edu.byu.cs.client.R;
 import edu.byu.cs.tweeter.client.cache.Cache;
 import edu.byu.cs.tweeter.client.presenter.RegisterPresenter;
-import edu.byu.cs.tweeter.client.view.AuthenticationView;
+import edu.byu.cs.tweeter.client.presenter.view.AuthenticationView;
 import edu.byu.cs.tweeter.client.view.main.MainActivity;
 import edu.byu.cs.tweeter.model.domain.User;
